Common fit definitions recieves the most significant changes:   Adds axis blob definition to simplify use of non-standard histos.   Modifies state loader helpers to allow state files to be split by    sample   Adds ND_FHC_OA off axis sample type to enum definition.   Adds a number of new binning and axis definitions that have been     useful in debugging fits.
MissingProtonFakeFata.h is tweaked to fix bugs in weight selection.
  Input histograms are updated and renamed to nu/nubar to avoid
    confusion that weights are beam-mode dependent when they are
    true species dependent.

Adds attempt to programatically retrieve if HESSE had to padd the
diagonal of the covariance matrix, but it doesn't seem to work.

Core/GenieWeightList.cxx adds FSILikeEAvailSmearing and gets a clang-format

Fixes bug in new flux systs that resulted in events on the RHS
of the near detector not getting flux weights.
  Updates flux syst file so that decomposed parameters tend to
   cause positive variations (+1 -> more events).