Adds latest ND covmat.
Tidies RunFitPoint arguments to reduce confusion.

GetPredictionInterps now guesses whether it needs to
add the sample name to the input file on its lonesome.

Adds export OMP_NUM_THREADS=1 to CAFAnaEnv.sh as latest version
of ROOT seems to enable omp by default in MINUIT2 with no
minimizer interface way of telling it to be single threaded
(which seems insane, but this fixes the apparent horrible
crashes in odd places that I was seeing.)