rework the BayesianMarginal classes so that they function with 'histogram' mode too
The BayesianMarginal classes were originally designed to determine the
breakpoints for various quantiles via the log prob values.  This is
great theoretically, but for some reason samples occasionally get funky
log probs from Stan, even though the sample density seems to still be
right.  These funky log probs screw up the other methods.

To work correctly with the 'histogram' method, we have to figure out
what the breakpoint is in the posterior probability determined from the
histogram rather than Stan's value.  This requires pulling it out of the
histogram itself.

This commit implements that.