Improved the implementation of the detector systematics. Previously, the ERec vs ETrue matrices could have more ETrue bins than ERec bins. This would creat an underdetermined system that was impossible to solve. The binning is now the same for ERec and ETrue in the ND and FD matrices.
BUG fix for the NDFVSliceCorrection. Now it corrects the POT exposure
properly for the off-axis bin width. This means the LC weights are the
correct sizes. Properly removed the horrible bidge jobs I previously
added.

Added scaling to the energy bin width in the PRISMPrediction app
as I don't like the look of it when using an alternative binning.